+++ VehicleRootingProblem/Visualization.cpp
@@ -11,7 +11,7 @@ using namespace System::Drawing;
 void DrawPaths(System::Drawing::Graphics^ graphics, ProblemSolution solution, int height, int width, int cur_time) {
 	int sumPoints = 0;
 	for (auto path : solution.Paths) {
-		sumPoints += path.size();
+		sumPoints += static_cast<int>(path.size());
 	}
 	if (!sumPoints) {
 		return;
@@ -91,8 +91,8 @@ void DrawPaths(System::Drawing::Graphics^ graphics, ProblemSolution solution, in
 
 			auto drawFont = gcnew Font("Arial", 12);
 			auto drawBrush = gcnew SolidBrush(Color::Black);
-			float x = pt.first - 1;
-			float y = pt.second + 5;
+			float x = static_cast<float>(pt.first - 1.0);
+			float y = static_cast<float>(pt.second + 5.0);
 			auto drawFormat = gcnew StringFormat();
 			graphics->DrawString(text, drawFont, drawBrush, x, y, drawFormat);
 		}
